requirements.txt
alembic==1.4.2
autopep8==2.0.1
blinker==1.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==6.0.0
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pipfile==0.0.2
pycodestyle==2.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
toml==0.10.1
tomli==2.0.1
validators==0.18.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
zipp==3.12.0
